backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic[email]
PyMuPDF
httpx
python-dotenv
sqlmodel
passlib[bcrypt]
python-jose
psycopg2-binary
sqlalchemy
alembic
pydantic-settings
pytest-asyncio
pytest-cov
pytest-mock
pytest-factory
pytest
jwt